Log version info in CDS messages, move debug method to debug file
