+++ pilot/pkg/proxy/envoy/v2/ads.go
@@ -23,10 +23,8 @@ import (
 	"sync"
 	"time"
 
-	adminapi "github.com/envoyproxy/go-control-plane/envoy/admin/v2alpha"
 	xdsapi "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	ads "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v2"
-	"github.com/gogo/protobuf/types"
 	"github.com/prometheus/client_golang/prometheus"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -238,64 +236,6 @@ type XdsConnection struct {
 	pushMutex sync.Mutex
 }
 
-// configDump converts the connection internal state into an Envoy Admin API config dump proto
-// It is used in debugging to create a consistent object for comparison between Envoy and Pilot outputs
-func (s *DiscoveryServer) configDump(conn *XdsConnection) (*adminapi.ConfigDump, error) {
-	configDump := &adminapi.ConfigDump{Configs: map[string]types.Any{}}
-
-	dynamicActiveClusters := []adminapi.ClustersConfigDump_DynamicCluster{}
-	clusters, err := s.generateRawClusters(conn, s.env.PushContext)
-	if err != nil {
-		return nil, err
-	}
-	for _, cs := range clusters {
-		dynamicActiveClusters = append(dynamicActiveClusters, adminapi.ClustersConfigDump_DynamicCluster{Cluster: cs})
-	}
-	clustersAny, err := types.MarshalAny(&adminapi.ClustersConfigDump{
-		VersionInfo:           versionInfo(),
-		DynamicActiveClusters: dynamicActiveClusters,
-	})
-	if err != nil {
-		return nil, err
-	}
-	configDump.Configs["clusters"] = *clustersAny
-
-	dynamicActiveListeners := []adminapi.ListenersConfigDump_DynamicListener{}
-	listeners, err := s.generateRawListeners(conn, s.env.PushContext)
-	if err != nil {
-		return nil, err
-	}
-	for _, cs := range listeners {
-		dynamicActiveListeners = append(dynamicActiveListeners, adminapi.ListenersConfigDump_DynamicListener{Listener: cs})
-	}
-	listenersAny, err := types.MarshalAny(&adminapi.ListenersConfigDump{
-		VersionInfo:            versionInfo(),
-		DynamicActiveListeners: dynamicActiveListeners,
-	})
-	if err != nil {
-		return nil, err
-	}
-	configDump.Configs["listeners"] = *listenersAny
-
-	routes, err := s.generateRawRoutes(conn, s.env.PushContext)
-	if err != nil {
-		return nil, err
-	}
-	if len(routes) > 0 {
-		dynamicRouteConfig := []adminapi.RoutesConfigDump_DynamicRouteConfig{}
-		for _, rs := range routes {
-			dynamicRouteConfig = append(dynamicRouteConfig, adminapi.RoutesConfigDump_DynamicRouteConfig{RouteConfig: rs})
-		}
-		routeConfigAny, err := types.MarshalAny(&adminapi.RoutesConfigDump{DynamicRouteConfigs: dynamicRouteConfig})
-		if err != nil {
-			return nil, err
-		}
-		configDump.Configs["routes"] = *routeConfigAny
-	}
-
-	return configDump, nil
-}
-
 // XdsEvent represents a config or registry event that results in a push.
 type XdsEvent struct {
 
@@ -434,7 +374,7 @@ func (s *DiscoveryServer) StreamAggregatedResources(stream ads.AggregatedDiscove
 				// soon as the CDS push is returned.
 				adsLog.Infof("ADS:CDS: REQ %v %s %v raw: %s", peerAddr, con.ConID, time.Since(t0), discReq.String())
 				con.CDSWatch = true
-				err := s.pushCds(con, s.env.PushContext)
+				err := s.pushCds(con, s.env.PushContext, versionInfo())
 				if err != nil {
 					return err
 				}
@@ -578,7 +518,7 @@ func (s *DiscoveryServer) pushAll(con *XdsConnection, pushEv *XdsEvent) error {
 	}
 
 	if con.CDSWatch {
-		err := s.pushCds(con, pushEv.push)
+		err := s.pushCds(con, pushEv.push, pushEv.version)
 		if err != nil {
 			return err
 		}
