+++ pilot/pkg/proxy/envoy/v2/debug.go
@@ -26,6 +26,7 @@ import (
 	"github.com/gogo/protobuf/jsonpb"
 
 	authn "istio.io/api/authentication/v1alpha1"
+	adminapi "github.com/envoyproxy/go-control-plane/envoy/admin/v2alpha"
 	meshconfig "istio.io/api/mesh/v1alpha1"
 	networking "istio.io/api/networking/v1alpha3"
 	"istio.io/istio/pilot/pkg/model"
@@ -33,6 +34,7 @@ import (
 	authn_plugin "istio.io/istio/pilot/pkg/networking/plugin/authn"
 	"istio.io/istio/pilot/pkg/serviceregistry"
 	"istio.io/istio/pilot/pkg/serviceregistry/aggregate"
+	"github.com/gogo/protobuf/types"
 )
 
 // memregistry is based on mock/discovery - it is used for testing and debugging v2.
@@ -780,3 +782,62 @@ func printRoutes(w io.Writer, c *XdsConnection) {
 		}
 	}
 }
+
+// configDump converts the connection internal state into an Envoy Admin API config dump proto
+// It is used in debugging to create a consistent object for comparison between Envoy and Pilot outputs
+func (s *DiscoveryServer) configDump(conn *XdsConnection) (*adminapi.ConfigDump, error) {
+	configDump := &adminapi.ConfigDump{Configs: map[string]types.Any{}}
+
+	dynamicActiveClusters := []adminapi.ClustersConfigDump_DynamicCluster{}
+	clusters, err := s.generateRawClusters(conn, s.env.PushContext)
+	if err != nil {
+		return nil, err
+	}
+	for _, cs := range clusters {
+		dynamicActiveClusters = append(dynamicActiveClusters, adminapi.ClustersConfigDump_DynamicCluster{Cluster: cs})
+	}
+	clustersAny, err := types.MarshalAny(&adminapi.ClustersConfigDump{
+		VersionInfo:           versionInfo(),
+		DynamicActiveClusters: dynamicActiveClusters,
+	})
+	if err != nil {
+		return nil, err
+	}
+	configDump.Configs["clusters"] = *clustersAny
+
+	dynamicActiveListeners := []adminapi.ListenersConfigDump_DynamicListener{}
+	listeners, err := s.generateRawListeners(conn, s.env.PushContext)
+	if err != nil {
+		return nil, err
+	}
+	for _, cs := range listeners {
+		dynamicActiveListeners = append(dynamicActiveListeners, adminapi.ListenersConfigDump_DynamicListener{Listener: cs})
+	}
+	listenersAny, err := types.MarshalAny(&adminapi.ListenersConfigDump{
+		VersionInfo:            versionInfo(),
+		DynamicActiveListeners: dynamicActiveListeners,
+	})
+	if err != nil {
+		return nil, err
+	}
+	configDump.Configs["listeners"] = *listenersAny
+
+	routes, err := s.generateRawRoutes(conn, s.env.PushContext)
+	if err != nil {
+		return nil, err
+	}
+	if len(routes) > 0 {
+		dynamicRouteConfig := []adminapi.RoutesConfigDump_DynamicRouteConfig{}
+		for _, rs := range routes {
+			dynamicRouteConfig = append(dynamicRouteConfig, adminapi.RoutesConfigDump_DynamicRouteConfig{RouteConfig: rs})
+		}
+		routeConfigAny, err := types.MarshalAny(&adminapi.RoutesConfigDump{DynamicRouteConfigs: dynamicRouteConfig})
+		if err != nil {
+			return nil, err
+		}
+		configDump.Configs["routes"] = *routeConfigAny
+	}
+
+	return configDump, nil
+}
+
